Move most ImGui-related files/code to an SohGui subfolder.
