+++ soh/soh/Enhancements/cosmetics/CosmeticsEditor.cpp
@@ -10,7 +10,7 @@
 #include <algorithm>
 #include <libultraship/libultraship.h>
 
-#include "soh/UIWidgets.hpp"
+#include "soh/SohGui/UIWidgets.hpp"
 #include "soh/OTRGlobals.h"
 #include "soh/ResourceManagerHelpers.h"
