+++ soh/soh/Enhancements/debugger/actorViewer.cpp
@@ -1,6 +1,6 @@
 #include "actorViewer.h"
 #include "../../util.h"
-#include "../../UIWidgets.hpp"
+#include "soh/SohGui/UIWidgets.hpp"
 #include "soh/ActorDB.h"
 #include "soh/Enhancements/game-interactor/GameInteractor.h"
 #include "soh/Enhancements/nametag.h"
