+++ soh/soh/Enhancements/gameplaystats.cpp
@@ -5,7 +5,7 @@
 #include "functions.h"
 #include "macros.h"
 #include "soh/cvar_prefixes.h"
-#include "../UIWidgets.hpp"
+#include "soh/SohGui/UIWidgets.hpp"
 #include "soh/util.h"
 
 #include <vector>
