+++ soh/soh/Enhancements/randomizer/randomizer_check_tracker.cpp
@@ -6,7 +6,7 @@
 #include "soh/cvar_prefixes.h"
 #include "soh/SaveManager.h"
 #include "soh/ResourceManagerHelpers.h"
-#include "soh/UIWidgets.hpp"
+#include "soh/SohGui/UIWidgets.hpp"
 #include "dungeon.h"
 #include "location_access.h"
