+++ soh/soh/Enhancements/resolution-editor/ResolutionEditor.cpp
@@ -2,7 +2,7 @@
 #include <imgui.h>
 #include <libultraship/libultraship.h>
 
-#include <soh/UIWidgets.hpp>
+#include "soh/SohGui/UIWidgets.hpp"
 #include <graphic/Fast3D/gfx_pc.h>
 #include "soh/OTRGlobals.h"
 #include "soh/cvar_prefixes.h"
