+++ soh/soh/SohGui/SohGui.cpp
@@ -25,15 +25,15 @@
 #include "include/global.h"
 #include "include/z64audio.h"
 #include "soh/SaveManager.h"
-#include "OTRGlobals.h"
+#include "soh/OTRGlobals.h"
 #include "soh/Enhancements/presets.h"
 #include "soh/resource/type/Skeleton.h"
 #include "libultraship/libultraship.h"
 
-#include "Enhancements/game-interactor/GameInteractor.h"
-#include "Enhancements/cosmetics/authenticGfxPatches.h"
-#include "Enhancements/resolution-editor/ResolutionEditor.h"
-#include "Enhancements/debugger/MessageViewer.h"
+#include "soh/Enhancements/game-interactor/GameInteractor.h"
+#include "soh/Enhancements/cosmetics/authenticGfxPatches.h"
+#include "soh/Enhancements/resolution-editor/ResolutionEditor.h"
+#include "soh/Enhancements/debugger/MessageViewer.h"
 #include "soh/Notification/Notification.h"
 #include "soh/Enhancements/TimeDisplay/TimeDisplay.h"
